Added splash screen
Close #90
Added splash screen to run before application and display name of the
printer.
It is really just a dummy screen, and not necessary for loading as most
imports have already completed because the splash screen requires wx.

"""Note: most additions to App.py in this commit are involved in testing
full screen mode.  I could not commit individual lines, so I included
them in this commit just to get this in on time.