THE GREAT PURGE
Master branch was extremely messed up from various probablylems with
merges, pushes and the subsequent pathetic attempts at fixing it. A ton
of development happened in the FullScreen branch, so the master branch
will be replaced by FullScreen. This "Purge" is deleting all of the
files, so the new ones can be imported.