Fixed coloring. Added events when data changed.
Some issues arose when I made changes for mac compatibility. Now the app
tests for OS and changes the way it draws accordingly.
When Data values are changed in a Dimension Editor or just Dynamic Data
Display object, an even is generated, so parents can listen for when a
value changes.