Added scale and offset operations.
In both part view and printing:
The user can specify the scale of the part relative to number in the STL
file (an associated unit library was added, but is not yet accessible).
Offset values for the part are shown will quickly update in the display
view and also affect the bitmap generation.